Pulling more game files

+++ dbt_cfb/models/t_team_d.sql
@@ -27,6 +27,6 @@ tmp
 
 {% if is_incremental() %}
 
-where team_id not in (select team_id from {{ this }})
+where team_id::int not in (select team_id from {{ this }})
 
 {% endif %}
